This function should return an int

+++ architecture/parameters/contextEncoderParameters.py
@@ -28,4 +28,4 @@ class ContextEncoderParameters(object):
         return self._fftWindowLength//2+1
 
     def _contextFrames(self):
-        return ((self._signalLength-self._gapLength)/2)/self._fftHopSize
+        return int(((self._signalLength-self._gapLength)/2)/self._fftHopSize)
